Modem code cleanup (leave only QSP implementation)
