+++ fmarkedlg.cpp
@@ -27,11 +27,7 @@
 #include <QString>
 
 #include "fmarkedlg.h"
-#ifdef Q_OS_WIN
-    #include "modem_win.h"
-#else
-    #include "modem_lin.h"
-#endif
+#include "modem.hpp"
 
 
 //#####################################################################################
