+++ nwt7mainclass.h
@@ -62,11 +62,7 @@
 #include <string.h>
 #include <qglobal.h>
 
-#ifdef Q_OS_WIN
-     #include "modem_win.h"
-#else
-     #include "modem_lin.h"
-#endif
+#include "modem.hpp"
 
 
 #include "sweepcurvewidget.h"
